+++ opennmt/__init__.py
@@ -1,6 +1,6 @@
 """OpenNMT module."""
 
-__version__ = "1.12.0"
+__version__ = "1.13.0"
 
 from opennmt import decoders
 from opennmt import encoders