+++ opennmt/models/transformer.py
@@ -99,11 +99,12 @@ class Transformer(SequenceToSequence):
                 "beta2": 0.998
             },
             "learning_rate": 2.0,
-            "decay_type": "noam_decay",
-            "decay_rate": self._num_units,
-            "decay_step_duration": max(1, 8 // num_devices),
-            "decay_steps": 8000,
-            "start_decay_steps": 0
+            "gradients_accum": max(1, 8 // num_devices),
+            "decay_type": "noam_decay_v2",
+            "decay_params": {
+                "model_dim": self._num_units,
+                "warmup_steps": 8000
+            }
         },
         "train": {
             "batch_size": 3072,